\docType{data}
\encoding{utf8}
\name{AgeGender}
\alias{AgeGender}
\title{Ages and genders}
\format{A table with 3 variables, to be queried with the SCB() function.
\describe{
\item{Municipality}{Also known as: region, Kommun.
Municipality the observation is from.}
\item{Age}{Also known as: ålder, Ålder.
A range of numeric values, taking values between 0 and 100.}
\item{Gender}{Also known as: kön, Kön.
A column of factor data. Its levels are: män (a.k.a.: Men, Män, Male.), kvinnor (a.k.a.: Women, Kvinnor, Female.).}
}}
\description{A table containing data on population, divided by age, gender, and municipality.}
\keyword{datasets}
