\docType{data}
\encoding{utf8}
\name{BirthRegion}
\alias{BirthRegion}
\title{Region of birth}
\format{A table with 4 variables, to be queried with the SCB() function.
\describe{
\item{Municipality}{Also known as: region, Kommun.
Municipality the observation is from.}
\item{Age}{Also known as: ålder, Ålder.
A range of numeric values, taking values between 12 and 25.}
\item{Gender}{Also known as: kön, Kön.
A column of factor data. Its levels are: män (a.k.a.: Men, Män, Male.), kvinnor (a.k.a.: Women, Kvinnor, Female.).}
\item{BirthRegion}{Also known as: födelseregion, Födelseregion.
A column of factor data. Its levels are: Född i Sverige (a.k.a.: Sverige, Born in Sweden, Sweden.), Utrikes född (a.k.a.: Inte Sverige, Born outside Sweden, Not Sweden.).}
}}
\description{A table containing data on country of birth (Sweden or not Sweden), divided by age and gender.}
\keywords{datasets}
