\docType{data}
\encoding{utf8}
\name{EducationLevel}
\alias{EducationLevel}
\title{Level of education}
\format{A table with 4 variables, to be queried with the SCB() function.
\describe{
\item{Municipality}{Also known as: region, Kommun.
Municipality the observation is from.}
\item{Age}{Also known as: ålder, Ålder.
A column of factor data. Its levels are: 16-24 år (a.k.a.: 16-24.), 25-34 år (a.k.a.: 25-34.), 35-44 år (a.k.a.: 35-44.), 45-54 år (a.k.a.: 45-54.), 55-64 år (a.k.a.: 55-64.), 65-74 år (a.k.a.: 65-74.), 75-84 år (a.k.a.: 75-84.), 85-94 år (a.k.a.: 85-94.), 95+ år (a.k.a.: 95+.).}
\item{Gender}{Also known as: kön, Kön.
A column of factor data. Its levels are: män (a.k.a.: Men, Män, Male.), kvinnor (a.k.a.: Women, Kvinnor, Female.).}
\item{EducationLevel}{Also known as: utbildningsnivå, Utbildningsnivå.
A column of factor data. Its levels are: eftergymnasial utbildning, 3 år eller mer (a.k.a.: Tertiary education, 3+ years.), eftergymnasial utbildning, mindre än 3 år (a.k.a.: Tertiary education, less than 3 years.), forskarutbildning (a.k.a.: Postgraduate education.), förgymnasial utbildning kortare än 9 år (a.k.a.: Primary education, less than 9 years.), förgymnasial utbildning, 9 (10) år (a.k.a.: Primary education, 9 (10) years.), gymnasial utbildning, 3 år (a.k.a.: Secondary education, 3 years.), gymnasial utbildning, högst 2 år (a.k.a.: Secondary education, at most 2 years.), uppgift om utbildningsnivå saknas (a.k.a.: No information, Uppgift saknas.).}
}}
\description{A table containing data on level of education, divided by age category and gender.}
\keywords{datasets}
