% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plotting.r
\name{example_forest_plot}
\alias{example_forest_plot}
\title{Create an example plot of forests in Sweden}
\usage{
example_forest_plot()
}
\description{
This function is an example of how one can integrate the \code{SCB()} function's handling of data with the
plotting utilities, to create a plot of how large a proportion of the area of each municipality is
forests.
}
\examples{
girafe(ggobj = example_forest_plot())

}
