% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOnMap.r
\name{exampleForestPlot}
\alias{exampleForestPlot}
\title{Create an example plot of forests in Sweden}
\usage{
exampleForestPlot()
}
\description{
This function is an example of how one can integrate the \code{SCB()} function's handling of data with the
plotting utilities, to create a plot of how large a proportion of the area of each municipality is
forests.
}
\examples{
library(ggplot2)
library(ggiraph)
girafe(ggobj = exampleForestPlot())

}
\seealso{
plotOnMap
}
