% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plotting.r
\name{fill_out_mapplotframe}
\alias{fill_out_mapplotframe}
\title{Fill out partial municipality data with NAs}
\usage{
fill_out_mapplotframe(dat)
}
\arguments{
\item{dat}{Data frame of partial data on municipalities}
}
\value{
A data frame with one row for each municipality
}
\description{
This function takes a data frame consisting of one column of municipality names and one column of data for each municipality,
and adds a row for each municipality that is not present, with NA in the data column. It also strips out any row which does
not have the name of a municipality in the first column.
}
\examples{
fill_out_mapplotframe(data.frame(municipality = c("Stockholm", "Lund", "Uppsala"), students = c(23000, 32000, 33500)))
}
