\docType{data}
\encoding{utf8}
\name{MunicipalIncomeStatements}
\alias{MunicipalIncomeStatements}
\title{Incomes of municipal governments}
\format{A table with 2 variables, to be queried with the SCB() function.
\describe{
\item{Municipality}{Also known as: region, Kommun.
Municipality the observation is from.}
\item{IncomeStatementPost}{Also known as: resultaträkningsposter, Resultaträkningspost.
A column of factor data. Its levels are: skatteintäkter (a.k.a.: Skatteintäkter, Tax income.), generella statsbidrag och utjämning (a.k.a.: SubsidiesAndEqualisation.).}
}}
\description{A table containing data on total tax income, and equalisation and subsidies, per municipality.}
\source{SCB, 2019}
\keyword{datasets}
