\docType{data}
\encoding{utf8}
\name{HouseholdSize}
\alias{HouseholdSize}
\title{Household size}
\format{A table with 2 variables, to be queried with the SCB() function.
\describe{
\item{Municipality}{Also known as: region, Kommun.
Municipality the observation is from.}
\item{HouseholdSize}{Also known as: hushållsstorlek, Hushållsstorlek.
A column of factor data. Its levels are: 1 person (a.k.a.: 1.), 2 personer (a.k.a.: 2, 2 people.), 3 personer (a.k.a.: 3, 3 people.), 4 personer (a.k.a.: 4, 4 people.), 5 personer (a.k.a.: 5, 5 people.), 6 personer (a.k.a.: 6, 6 people.), 7+ personer (a.k.a.: 7+, 7+ people.), uppgift saknas (a.k.a.: Information missing.).}
}}
\description{A table containing data on how many households of each given size there are in a municipality.}
\keywords{datasets}
