\docType{data}
\encoding{utf8}
\name{HousingForm}
\alias{HousingForm}
\title{Housing form}
\format{A table with 3 variables, to be queried with the SCB() function.
\describe{
\item{Municipality}{Also known as: region, Kommun.
Municipality the observation is from.}
\item{HousingForm}{Also known as: boendeform, Boendeform.
A column of factor data. Its levels are: flerbostadshus, bostadsrätt (a.k.a.: Multi-home house, condo.), flerbostadshus, hyresrätt (a.k.a.: Multi-home house, rented.), småhus, bostadsrätt (a.k.a.: Small house, condo.), småhus, hyresrätt (a.k.a.: Small house, rented.), småhus, äganderätt (a.k.a.: Small house, owned.), specialbostad (a.k.a.: Special housing.), uppgift saknas (a.k.a.: No information.), övrigt boende (a.k.a.: Other, Övrigt.).}
\item{HouseholdType}{Also known as: hushållstyp, Hushållstyp.
A column of factor data. Its levels are: ensamstående kvinnor -64 år utan barn (a.k.a.: Single women up to 64 without children, Single F up to 64 w.o. children.), ensamstående kvinnor 65- år utan barn (a.k.a.: Single women above 64 without children, Single F above 64 w.o. children.), ensamstående kvinnor med barn 0-24 år (a.k.a.: Single women with children, Single F w. childen.), ensamstående män -64 år utan barn (a.k.a.: Single men up to 64 without children, Single M up to 64 w.o. childen.), ensamstående män 65- år utan barn (a.k.a.: Single men above 64 without children, Single M above 64 w.o. children.), ensamstående män med barn 0-24 år (a.k.a.: Single men with children, Single M with children.), sammanboende med barn 0-24 år, 1 barn (a.k.a.: Cohabiting with one child.), sammanboende med barn 0-24 år, 2 barn (a.k.a.: Cohabiting with two children.), sammanboende med barn 0-24 år, 3+  barn (a.k.a.: Cohabiting with 3+ children.), sammanboende utan barn (a.k.a.: Cohabiting without children.), övriga hushåll med barn 0-24 år (a.k.a.: Other households with children.), övriga hushåll utan barn 0-24 år (a.k.a.: Other households without childen.).}
}}
\description{A table containing data on how many households of a certain form (e.g. single male above 64, woman with child) live in a certain type of housing (e.g. condo, owned house).}
\source{SCB, 2019}
\keyword{datasets}
