\docType{data}
\encoding{utf8}
\name{Incomes}
\alias{Incomes}
\title{Incomes of individuals}
\format{A table with 4 variables, to be queried with the SCB() function.
\describe{
\item{Municipality}{Also known as: region, Kommun.
Municipality the observation is from.}
\item{Age}{Also known as: ålder, Ålder.
A column of factor data. Its levels are: 16-19 år (a.k.a.: 16-19, 16-19 y.o..), 20-24 år (a.k.a.: 20-24, 20-24 y.o..), totalt 16+ år (a.k.a.: Total 16+.), totalt 20+ år (a.k.a.: Total 20+.).}
\item{Gender}{Also known as: kön, Kön.
A column of factor data. Its levels are: män (a.k.a.: Men, Män, Male.), kvinnor (a.k.a.: Women, Kvinnor, Female.), totalt (a.k.a.: Totalt, Total.).}
\item{MeanOrMedian}{Also known as: MedTyp, MedelvärdeEllerMedian.
A column of factor data. Its levels are: Medelvärde (a.k.a.: Mean.), Median.}
}}
\description{A table containing data on mean and median income, divided by gender and some age categories.}
\keywords{datasets}
