\docType{data}
\encoding{utf8}
\name{LandUse}
\alias{LandUse}
\title{Use of land}
\format{A table with 2 variables, to be queried with the SCB() function.
\describe{
\item{Municipality}{Also known as: region, Kommun.
Municipality the observation is from.}
\item{LandUseType}{Also known as: markanvändningsklass, Markanvändningsklass.
A column of factor data. Its levels are: bebyggd mark och tillhörande mark (a.k.a.: Bebyggd, Built-upon and adherent land, Buildings.), betesmark (a.k.a.: Betesmark, Pasture.), golfbanor och skidpistar (a.k.a.: Golfbanor och skidpistar, Golf courses and ski pistes.), skogsmark, improduktiv (a.k.a.: Improduktiv skogsmark, Unproductive forest.), skogsmark, produktiv (a.k.a.: Produktiv skogsmark, Productive forest.), total landareal (a.k.a.: Total area.), total skogsmark (a.k.a.: Skogsmark, Forest, Total forest.), täkter och gruvområden (a.k.a.: Täkter och gruvområden, Borrows and mines.), vatten (a.k.a.: Vatten, Water.), åkermark (a.k.a.: Åkermark, Farmland.), övrig mark (a.k.a.: Övrig mark, Other land.).}
}}
\description{A table containing data on land use in each municipality.}
\keyword{datasets}
