\docType{data}
\encoding{utf8}
\name{MaritalStatus}
\alias{MaritalStatus}
\title{Marital status}
\format{A table with 4 variables, to be queried with the SCB() function.
\describe{
\item{Municipality}{Also known as: region, Kommun.
Municipality the observation is from.}
\item{Age}{Also known as: ålder, Ålder.
A range of numeric values, taking values between 0 and 100.}
\item{Gender}{Also known as: kön, Kön.
A column of factor data. Its levels are: män (a.k.a.: Men, Män, Male.), kvinnor (a.k.a.: Women, Kvinnor, Female.).}
\item{MaritalStatus}{Also known as: civilstånd, Civilstånd.
A column of factor data. Its levels are: ogifta (a.k.a.: Unmarried, Never married, Ogifta, Aldrig gifta.), gifta (a.k.a.: Married, Gifta.), skilda (a.k.a.: Divorced, Skilda.), änkor/änklingar (a.k.a.: WidowErs, Änk(or/lingar), Spouse deceased.).}
}}
\description{A table containing data on marital status, divided by age, gender, and municipality. Note that in the column Age, the value 100 actually represents all ages above 99.}
\keyword{datasets}
