% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plotting.r
\name{plot_on_map}
\alias{plot_on_map}
\title{Create interactive plot of data on map of Sweden}
\usage{
plot_on_map(
  dat,
  tooltips = NA,
  mainTitle = NA,
  subTitle = NA,
  legendTitle = ""
)
}
\arguments{
\item{dat}{Either a data frame of two columns, the first of which contains names of municipalities and the second
the data, or a named list, where the names are names of municipalities. Allowed to be partial data,
municipalities without data are set to NA.}

\item{tooltips}{A data frame with two columns, the first of which contains the name of a municipality and the second
the tooltip that should be shown for that municipality. If omitted, tooltips are generated automatically from the data.}

\item{mainTitle}{Main title of the plot, displayed at the top. If omitted, plot has no title.}

\item{subTitle}{Subtitle of the plot, displayed right below the title. If omitted, plot has no subtitle.}

\item{legendTitle}{Title of the legend of the plot. If omitted, legend has no title - so don't omit this.}
}
\value{
A ggplot object that can be turned into an interactive graphic with the girafe function, see the example.
}
\description{
This function takes in data on Swedish municipalities, and creates a plot of said data on an interactive
map of Sweden.
}
\examples{
exampledata <- data.frame(municipality = c("Stockholm", "Lund", "Uppsala"), students = c("some", "many", "loads"))
girafe(ggobj = plot_on_map(exampledata))

}
