\docType{data}
\encoding{utf8}
\name{TaxRates}
\alias{TaxRates}
\title{Tax rates}
\format{A table with 2 variables, to be queried with the SCB() function.
\describe{
\item{Municipality}{Also known as: region, Kommun.
Municipality the observation is from.}
\item{MunicipalOrRegionalTax}{Also known as: Nivå, KommunEllerRegionalskatt.
A column of factor data. Its levels are: Kommun (a.k.a.: Municipal.), Region (a.k.a.: Regional.).}
}}
\description{A table containing data on municipal and regional tax rates.}
\keywords{datasets}
