\docType{data}
\encoding{utf8}
\name{HighSchoolEligibility}
\alias{HighSchoolEligibility}
\title{High school eligibility}
\format{A table with 4 variables, to be queried with the SCB() function.
\describe{
\item{Municipality}{Also known as: Kommun, Kommun.
Municipality the observation is from.}
\item{Gender}{Also known as: Kön, Kön.
A column of factor data. Its levels are: Pojkar (a.k.a.: Men, Boys, Män, Male.), Flickor (a.k.a.: Girls, Women, Kvinnor, Female.), Samtliga (a.k.a.: Totalt, Total.).}
\item{SchoolOrganizer}{Also known as: Huvudman, Huvudman.
A column of factor data. Its levels are: Kommunal (a.k.a.: Municipal.), Enskild (a.k.a.: Private.), Samtliga (a.k.a.: Totalt, Total.).}
\item{HighSchoolEligibility}{Also known as: Behörighet, HSEligibility, Gymnasiebehörighet.
A column of factor data. Its levels are: Gymnasiebehöriga (a.k.a.: True, Behöriga, Eligible.), Gymnasieobehöriga (a.k.a.: False, Obehöriga, Ineligible.).}
}}
\description{A table containing data on how many students graduated ninth grade, and what proportion were eligible for high school, divided by gender and organizational form (municipal or private).}
\keywords{datasets}
