% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCB_fetcher.r
\name{SCB}
\alias{SCB}
\title{Fetch data from SCB tables}
\usage{
SCB(
  Municipality = NA,
  Age = NA,
  Gender = NA,
  BornInCountry = NA,
  HouseholdType = NA,
  HousingForm = NA,
  TotalHouseholds = NA,
  MaritalStatus = NA,
  HouseholdSize = NA,
  MunicipalTax = NA,
  RegionTax = NA,
  LandUseClass = NA,
  TotalArea = NA,
  MeanIncome = NA,
  MedianIncome = NA,
  TaxIncome = NA,
  SubsidiesAndEqualisation = NA,
  Education = NA,
  TotalPopulation = NA,
  TotalStudents = NA,
  HighSchoolEligible = NA,
  SchoolOrganiser = NA
)
}
\arguments{
\item{Municipality}{Municipality - general parameter, occurs in every table.}

\item{Age}{Age - general parameter, occurs in all tables except land use, tax rate, and financial result}

\item{Gender}{Gender - general parameter, occurs in all tables except land use, tax rate, and financial result}

\item{BornInCountry}{Count people born in Sweden. Set to TRUE to count people born in Sweden and FALSE to count foreign-born people. Table also contains Age and Gender.}

\item{HouseholdType}{Family structure of a household. Specific parameter. Table also contains HousingForm.}

\item{HousingForm}{Housing type of a household, e.g. rented apartment or house. Table also contains HouseholdType.}

\item{SchoolOrganiser}{}
}
\value{
A number fetched from the requested table.
}
\description{
This function reads data from the collected SCB tables. It infers which table to look at based on which
parameters are set and which are not. So there are two types of parameter -- general ones which occur
in several tables (Municipality, Gender, Age) and specific ones which occur only in one table. There are
also specific parameters to fetch data from a table
The function looks for a specific parameter set, and then looks at that table. If any parameter contained
in the table -- general or specific -- is not set, it will sum over all possible values of that parameter.
So, for example, \code{SCB(Municipality="Stockholm", Age=19, Gender="män", TotalPopulation=TRUE)} will return the
number of 19-year-old men in Stockholm, while \code{SCB(Municipality="Stockholm", Gender="män", TotalPopulation=TRUE)}
will return the total number of men, since it sums over all possible values of "Age". Similarly, passing a list of
parameters will give the sum over all elements of the list, so \code{SCB(Municipality="Stockholm", Age=c(13:19), Gender="män", TotalPopulation=TRUE)}
would count the number of teenagers.
}
\examples{
SCB(Municipality = "Stockholm", Age=24, Gender="kvinnor", MaritalStatus = "skilda")
SCB(Municipality = "Stockholm", LandUseClass = "total skogsmark")

}
